+++ apback/src/main/java/com/apback/repository/ProyectoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.apback.model.Proyecto;
 
 @Repository
-public interface ProyectoRepository extends JpaRepository<Proyecto, Long> {
+public interface ProyectoRepository extends JpaRepository<Proyecto, Integer> {
 
 }
